requirements.txt
cython==0.29.37
pyyaml==6.0.2
argparse==1.4.0
onnxruntime==1.19.0
tqdm==4.66.5
nemo_toolkit @ git+https://github.com/AI4Bharat/NeMo@nemo-v2
transformers==4.40.0
huggingface_hub==0.23
pytorch-lightning==2.4.0
hydra-core==1.3.2
librosa==0.10.2.post1
sentencepiece==0.2.0
pandas==2.2.2
lhotse==1.27.0
editdistance==0.8.1
jiwer==3.0.4
pyannote.audio==3.3.1
webdataset==0.2.100
datasets==2.21.0
IPython==8.27.0
joblib==1.4.2
